Implement Phase 0/1 foundations: async session API, decimals, generator.
- Add TMoney/TRate and seeded RNG in tpcc/domain
- Replace transactions/session.h with async TFuture API and TSemanticOp variant
- Add tpcc/generator population helpers and wire them into PostgreSQL ImportSync
- Add ILoadAdapter skeleton; fix c_ytd_payment to DECIMAL
- Unit tests for money, RNG, generator, and session types
- Document known defects and module status in dependencies.md

+++ tpcc/dbms/pgsql/init.cpp
@@ -96,7 +96,7 @@ CREATE TABLE customer (
     c_first        varchar(16)    NOT NULL,
     c_credit_lim   decimal(12, 2) NOT NULL,
     c_balance      decimal(12, 2) NOT NULL,
-    c_ytd_payment  float          NOT NULL,
+    c_ytd_payment  decimal(12, 2) NOT NULL,
     c_payment_cnt  int            NOT NULL,
     c_delivery_cnt int            NOT NULL,
     c_street_1     varchar(20)    NOT NULL,
